Introduce achievements and rework user route
Achievements can now be created, updated, and all that fun stuff.
In addition, the user route was modified to support built in filtering
by group and what achievements they accomplished.

Moreover, the route was updated so that all Django users are queryable
(even if they don't share a group). However, the permissions were updatee to
block queries for groups and achievements that are unknown to the user